+++ PBL2_9. taldea/PBl2_9. taldea/imagen.h
@@ -4,7 +4,7 @@
 #define MAX_IMG 100
 
 int  irudiaKargatu(char* fileName);
-void  irudiaMugitu(int numImg, double x, double y);
+void  irudiaMugitu(int numImg, float x, float y);
 void irudiakMarraztu(void);
 void irudiaKendu(int id);
